requirements.txt
webvtt_py==0.4.6
yt_dlp==2023.11.16
